feat: move `accept` to trait
